+++ src/main/java/com/moayong/api/domain/leaguemember/dto/response/MemberRankingResponse.java
@@ -10,7 +10,7 @@ public record MemberRankingResponse (
         String nickname,
         PromotionStatus promotionStatus,
         Integer rank,
-        Float rate,
+        Integer rate,
         Integer totalScore
 ){
 }
